+++ src/Core/ContractTypeBuilders/NodeBuilder.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace MissionControl.ContractTypeBuilders {
+  public interface NodeBuilder {
+    void Build();
+  }
+}
